runningWithTheVisionConference/serializers.py
from rest_framework import serializers
from .models import Registration, Moment

class RegistrationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Registration
        fields = "__all__"


class MomentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Moment
        fields = "__all__"
