requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
deprecation==2.1.0
Django==5.1.2
django-cors-headers==4.7.0
django-countries==7.6.1
django-environ==0.11.2
django-filter==25.1
django-guardian==2.4.0
django-js-asset==3.1.2
django-localflavor==4.0
django-mptt==0.16.0
django-oauth-toolkit==3.0.1
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-request==1.6.3
django-requests==0.0.2
django-rest-framework==0.1.0
django-restframework==0.0.1
django-tinymce==4.1.0
djangorestframework==3.15.2
forex-python==1.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jwcrypto==1.5.6
kombu==5.4.2
lxml==5.3.1
Markdown==3.7
msgpack==1.1.1
oauthlib==3.3.1
olefile==0.47
outcome==1.3.0.post0
packaging==25.0
phonenumbers==8.13.55
pillow==11.0.0
postgrest==1.1.1
prompt_toolkit==3.0.48
psycopg2-binary
pycparser==2.22
pydantic==2.11.8
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-stdnum==1.20
pytz==2024.2
realtime==2.7.0
redis==5.2.0
requests==2.32.3
selenium==4.29.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.9.0
wsproto==1.2.0
